backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-dotenv
redis
python-jose[cryptography]
passlib
pytest
httpx
google-cloud-speech
google-cloud-storage
